code/2-Node-Type-B-Rpi0-CODE/EpalSlaveRpi/controller/requirements.txt
python-dotenv~=0.15.0
paho-mqtt~=1.5.1
